refactor(vr): fold ImGuiVRHelper glue into the VR feature
The migration left the VR feature mostly empty and put the helper-client glue in a
parallel ImGuiVRHelperClient namespace (src/ root) plus a StatusHUD class (Menu/),
outside the feature boundary the old VR code respected. Fold both onto the VR
feature instead — no new classes:

- ImGuiVRHelperClient + StatusHUD become VR:: methods in Features/VR/HelperClient.cpp
  (ConnectHelper/UpdateHelper/RenderHelperToPanel/RenderStatusHud/FeedHelperEvent/
  IsHelperRegistered/HelperRequestsRender/DrawHelperBindingsTable), declared in the
  VR feature's previously-empty "VR-SPECIFIC PUBLIC API" section.
- Connect now runs from the existing VR::PostPostLoad (feature lifecycle), so the
  explicit ImGuiVRHelperClient::Init() call in XSEPlugin is gone.
- DrawSettings calls its own DrawHelperBindingsTable; Menu/OverlayRenderer call
  globals::features::vr.* for the two cross-cutting touch-points (render + input).
- Delete ImGuiVRHelperClient.{cpp,h} and StatusHUD.{cpp,h}.

No behavior change; builds clean. Pure relocation onto the feature boundary.
